@@ -3,14 +3,15 @@ BatchProcessor_2 for sequential batch processing of prompts through LLMClient.
 Adapted from old_code/batch_processor.py with enhancements for dirs, formats, etc.
 Compatible with src/generation/ module.
 """
-import os
+
+import csv
 import json
 import time
-import csv
 from pathlib import Path
-from typing import List, Optional, Dict
+
 from rich.console import Console
 from rich.progress import track
+
 from .generator import LLMClient
 
 
@@ -22,16 +23,16 @@ class BatchProcessor_2:
         self.console = Console()
         self.client = LLMClient()
 
-    def load_prompts(self, input_path: str) -> List[str]:
+    def load_prompts(self, input_path: str) -> list[str]:
         """Load prompts from file or directory of .txt files."""
         path = Path(input_path).resolve()
         if not path.exists():
             self.console.print(f"[red]❌ Input path not found: {input_path}[/red]")
             raise FileNotFoundError(f"Input path not found: {input_path}")
 
-        prompts: List[str] = []
-        if path.is_file() and path.suffix.lower() == '.txt':
-            with open(path, "r", encoding="utf-8") as f:
+        prompts: list[str] = []
+        if path.is_file() and path.suffix.lower() == ".txt":
+            with Path(path).open("r", encoding="utf-8") as f:
                 file_prompts = [line.strip() for line in f if line.strip()]
                 prompts.extend(file_prompts)
                 self.console.print(f"[green]✅ Loaded {len(file_prompts)} prompts from {path.name}[/green]")
@@ -41,7 +42,7 @@ class BatchProcessor_2:
                 self.console.print("[yellow]⚠️ No .txt files found in directory.[/yellow]")
                 return prompts
             for txt_file in sorted(txt_files):
-                with open(txt_file, "r", encoding="utf-8") as f:
+                with Path(txt_file).open("r", encoding="utf-8") as f:
                     file_prompts = [line.strip() for line in f if line.strip()]
                     prompts.extend(file_prompts)
                     self.console.print(f"[green]✅ Loaded {len(file_prompts)} from {txt_file.name}[/green]")
@@ -52,16 +53,16 @@ class BatchProcessor_2:
 
     def process_prompts(
         self,
-        prompts: List[str],
-        output_file: Optional[str] = None,
+        prompts: list[str],
+        output_file: str | None = None,
         output_format: str = "jsonl",
-    ) -> List[Dict[str, str]]:
+    ) -> list[dict[str, str]]:
         """Process prompts sequentially, print results, optionally save."""
         if not prompts:
             self.console.print("[yellow]⚠️ No prompts to process.[/yellow]")
             return []
 
-        results: List[Dict[str, str]] = []
+        results: list[dict[str, str]] = []
         for idx, prompt in enumerate(track(prompts, description="Processing prompts...")):
             self.console.print(f"\n[bold yellow]Prompt {idx + 1}/{len(prompts)}:[/bold yellow]")
             self.console.print(f"[italic cyan]{prompt}[/italic cyan]\n")
@@ -71,7 +72,7 @@ class BatchProcessor_2:
                 self.console.print(f"[bold green]Response:[/bold green]\n{content}\n")
                 result = {"prompt": prompt, "response": content}
             except Exception as e:
-                content = f"ERROR: {str(e)}"
+                content = f"ERROR: {e!s}"
                 self.console.print(f"[bold red]{content}[/bold red]\n")
                 result = {"prompt": prompt, "response": content}
 
@@ -90,22 +91,22 @@ class BatchProcessor_2:
         self.console.print(f"[bold green]✅ Completed {len(results)} prompts.[/bold green]")
         return results
 
-    def _append_jsonl(self, output_file: str, result: Dict[str, str]) -> None:
+    def _append_jsonl(self, output_file: str, result: dict[str, str]) -> None:
         """Append single result to JSONL file."""
         Path(output_file).parent.mkdir(parents=True, exist_ok=True)
-        with open(output_file, "a", encoding="utf-8") as f:
+        with Path(output_file).open("a", encoding="utf-8") as f:
             json.dump(result, f)
             f.write("\n")
 
-    def _save_batch(self, output_file: str, results: List[Dict[str, str]], fmt: str) -> None:
+    def _save_batch(self, output_file: str, results: list[dict[str, str]], fmt: str) -> None:
         """Save batch results in specified format."""
         Path(output_file).parent.mkdir(parents=True, exist_ok=True)
         try:
             if fmt == "json":
-                with open(output_file, "w", encoding="utf-8") as f:
+                with Path(output_file).open("w", encoding="utf-8") as f:
                     json.dump(results, f, indent=2)
             elif fmt == "csv":
-                with open(output_file, "w", encoding="utf-8", newline="") as f:
+                with Path(output_file).open("w", encoding="utf-8", newline="") as f:
                     writer = csv.DictWriter(f, fieldnames=["prompt", "response"])
                     writer.writeheader()
                     writer.writerows(results)
