@@ -2,22 +2,24 @@
 Prompt construction and context building for generation
 """
 
-from abc import ABC, abstractmethod
-from typing import List, Dict, Any, Optional
 import os
-import sys
+from abc import ABC
+from abc import abstractmethod
 from textwrap import dedent
+from typing import Any
+from typing import Optional
 
 
 class ContextPromptBuilder(ABC):
     @abstractmethod
-    def build_prompt(self, chunk: Dict[str, Any], symbol_index: Optional['SymbolIndex'] = None) -> str:
+    def build_prompt(self, chunk: dict[str, Any], symbol_index: Optional["SymbolIndex"] = None) -> str:
         pass
 
 
 class SymbolIndex:
     """Enhanced symbol index for lookup with improved info extraction"""
-    def __init__(self, data: Optional[Dict[str, Any]] = None):
+
+    def __init__(self, data: dict[str, Any] | None = None):
         self.data = data or {}
 
     def get_info(self, symbol: str) -> str:
@@ -37,18 +39,17 @@ class SymbolIndex:
         # Enhanced info based on symbol kind
         if kind == "class":
             return f"Class defined in {short_file}:{line}"
-        elif kind == "function":
+        if kind == "function":
             return f"Function defined in {short_file}:{line}"
-        elif kind == "method":
+        if kind == "method":
             return f"Method defined in {short_file}:{line}"
-        else:
-            return f"{kind} defined in {short_file}:{line}"
+        return f"{kind} defined in {short_file}:{line}"
 
 
 class MultiLanguageContextPrompt(ContextPromptBuilder):
     """Enhanced prompt builder that handles multiple languages and chunk types"""
 
-    def build_prompt(self, chunk: Dict[str, Any], symbol_index: Optional[SymbolIndex] = None) -> str:
+    def build_prompt(self, chunk: dict[str, Any], symbol_index: SymbolIndex | None = None) -> str:
         # Build dependency context
         deps_info = self._build_dependency_context(chunk, symbol_index)
 
@@ -56,51 +57,51 @@ class MultiLanguageContextPrompt(ContextPromptBuilder):
         domain_context = self._get_domain_specific_context(chunk)
 
         # Build language-specific instructions
-        language_instructions = self._get_language_instructions(chunk.get('language', 'unknown'))
+        language_instructions = self._get_language_instructions(chunk.get("language", "unknown"))
 
         # Build chunk-type specific context
         chunk_type_context = self._get_chunk_type_context(chunk)
 
-        prompt = dedent(f"""You are an expert software developer specializing in {chunk.get('language', 'unknown')} and web development.
+        prompt = dedent(f"""You are an expert software developer specializing in {chunk.get("language", "unknown")} and web development.
         Summarize the following code chunk in one clear, concise sentence.
         Focus on its purpose, behavior, and role in the application.
 
         {domain_context}
 
-        File: {chunk.get('file_path', 'unknown')}
-        Type: {chunk.get('type', 'unknown')}
-        Name: {chunk.get('name', 'unknown')}
-        Qualified Name: {chunk.get('qualified_name', 'unknown')}
-        Location: Lines {chunk.get('location', {}).get('start_line', '?')}-{chunk.get('location', {}).get('end_line', '?')}
+        File: {chunk.get("file_path", "unknown")}
+        Type: {chunk.get("type", "unknown")}
+        Name: {chunk.get("name", "unknown")}
+        Qualified Name: {chunk.get("qualified_name", "unknown")}
+        Location: Lines {chunk.get("location", {}).get("start_line", "?")}-{chunk.get("location", {}).get("end_line", "?")}
 
         {chunk_type_context}
 
-        Dependencies: {', '.join(chunk.get('dependencies', [])) or 'None'}
+        Dependencies: {", ".join(chunk.get("dependencies", [])) or "None"}
         {deps_info}
 
         {language_instructions}
 
         Code:
-        ```{chunk.get('language', 'unknown')}
-        {chunk.get('code', '').strip()}
+        ```{chunk.get("language", "unknown")}
+        {chunk.get("code", "").strip()}
         ```
 
         Summary (one sentence, no markdown, no prefix):
         """).strip()
         return prompt
 
-    def _build_dependency_context(self, chunk: Dict[str, Any], symbol_index: Optional[SymbolIndex]) -> str:
+    def _build_dependency_context(self, chunk: dict[str, Any], symbol_index: SymbolIndex | None) -> str:
         """Build context about dependencies and relationships"""
         deps_info = []
         if symbol_index:
-            for dep in chunk.get('dependencies', [])[:5]:  # Limit to top 5 for brevity
+            for dep in chunk.get("dependencies", [])[:5]:  # Limit to top 5 for brevity
                 info = symbol_index.get_info(dep)
                 if info != "No additional info":
                     deps_info.append(f"- {dep}: {info}")
 
         # Add relationship context
         relationship_context = []
-        relationships = chunk.get('relationships', {})
+        relationships = chunk.get("relationships", {})
         if relationships:
             imports = relationships.get("imports", [])
             if imports:
@@ -122,26 +123,26 @@ class MultiLanguageContextPrompt(ContextPromptBuilder):
 
         return "\n".join(context_parts) if context_parts else "No additional dependency or relationship info."
 
-    def _get_domain_specific_context(self, chunk: Dict[str, Any]) -> str:
+    def _get_domain_specific_context(self, chunk: dict[str, Any]) -> str:
         """Get domain-specific context based on file path and content"""
-        context = chunk.get('context', {})
+        context = chunk.get("context", {})
         domain_ctx = context.get("domain_context", "General code")
         module_ctx = context.get("module_context", "Unknown module")
         project_ctx = context.get("project_context", "Project codebase")
 
         # Enhanced domain context based on file path
-        file_path = chunk.get('file_path', '').lower()
-        if any(term in file_path for term in ['admin', 'modeladmin']):
+        file_path = chunk.get("file_path", "").lower()
+        if any(term in file_path for term in ["admin", "modeladmin"]):
             domain_ctx = "Django admin configuration and model management"
-        elif any(term in file_path for term in ['model', 'schema']):
+        elif any(term in file_path for term in ["model", "schema"]):
             domain_ctx = "Data models and database schema"
-        elif any(term in file_path for term in ['view', 'controller']):
+        elif any(term in file_path for term in ["view", "controller"]):
             domain_ctx = "Application logic and request handling"
-        elif any(term in file_path for term in ['template', 'html']):
+        elif any(term in file_path for term in ["template", "html"]):
             domain_ctx = "User interface templates and presentation"
-        elif any(term in file_path for term in ['static', 'css']):
+        elif any(term in file_path for term in ["static", "css"]):
             domain_ctx = "Styling and user interface design"
-        elif any(term in file_path for term in ['static', 'js']):
+        elif any(term in file_path for term in ["static", "js"]):
             domain_ctx = "Client-side functionality and interactivity"
 
         return f"Module: {module_ctx}\nDomain: {domain_ctx}\nProject: {project_ctx}"
@@ -149,42 +150,42 @@ class MultiLanguageContextPrompt(ContextPromptBuilder):
     def _get_language_instructions(self, language: str) -> str:
         """Get language-specific instructions for the LLM"""
         instructions = {
-            'python': "Focus on Python-specific patterns, Django conventions if applicable, and the object's role in the application.",
-            'javascript': "Focus on JavaScript patterns, DOM manipulation if applicable, and the function's role in client-side logic.",
-            'html': "Focus on the template's structure, included components, and its role in the page layout and user interface.",
-            'css': "Focus on the styling rules, layout impact, and visual design role in the application.",
-            'java': "Focus on Java patterns, object-oriented design, and the class/method's role in the application architecture.",
-            'cpp': "Focus on C++ patterns, memory management considerations, and performance characteristics.",
-            'go': "Focus on Go patterns, concurrency if applicable, and the function's role in the system.",
-            'rust': "Focus on Rust patterns, ownership system, and safety characteristics."
+            "python": "Focus on Python-specific patterns, Django conventions if applicable, and the object's role in the application.",
+            "javascript": "Focus on JavaScript patterns, DOM manipulation if applicable, and the function's role in client-side logic.",
+            "html": "Focus on the template's structure, included components, and its role in the page layout and user interface.",
+            "css": "Focus on the styling rules, layout impact, and visual design role in the application.",
+            "java": "Focus on Java patterns, object-oriented design, and the class/method's role in the application architecture.",
+            "cpp": "Focus on C++ patterns, memory management considerations, and performance characteristics.",
+            "go": "Focus on Go patterns, concurrency if applicable, and the function's role in the system.",
+            "rust": "Focus on Rust patterns, ownership system, and safety characteristics.",
         }
         return instructions.get(language, "Focus on the code's purpose and behavior in the application.")
 
-    def _get_chunk_type_context(self, chunk: Dict[str, Any]) -> str:
+    def _get_chunk_type_context(self, chunk: dict[str, Any]) -> str:
         """Get context specific to the chunk type"""
-        chunk_type = chunk.get('type', 'unknown')
+        chunk_type = chunk.get("type", "unknown")
         type_contexts = {
-            'class': "This is a class definition. Describe its responsibility and how it might be used.",
-            'function': "This is a function. Describe what it does and its input/output behavior.",
-            'method': "This is a method within a class. Describe its specific role and how it modifies object state.",
-            'html_file': "This is a complete HTML template. Describe its overall structure and purpose in the UI.",
-            'html_element': "This is an HTML element or component. Describe its role in the page layout.",
-            'css_rule': "This is a CSS rule. Describe its styling purpose and visual impact."
+            "class": "This is a class definition. Describe its responsibility and how it might be used.",
+            "function": "This is a function. Describe what it does and its input/output behavior.",
+            "method": "This is a method within a class. Describe its specific role and how it modifies object state.",
+            "html_file": "This is a complete HTML template. Describe its overall structure and purpose in the UI.",
+            "html_element": "This is an HTML element or component. Describe its role in the page layout.",
+            "css_rule": "This is a CSS rule. Describe its styling purpose and visual impact.",
         }
 
         base_context = type_contexts.get(chunk_type, "Describe the purpose and behavior of this code.")
 
         # Add metadata-specific context
         metadata_context = []
-        metadata = chunk.get('metadata', {})
+        metadata = chunk.get("metadata", {})
         if metadata:
-            if metadata.get('decorators'):
+            if metadata.get("decorators"):
                 metadata_context.append(f"Decorators: {', '.join(metadata['decorators'])}")
-            if metadata.get('base_classes'):
+            if metadata.get("base_classes"):
                 metadata_context.append(f"Inherits from: {', '.join(metadata['base_classes'])}")
-            if metadata.get('is_abstract'):
+            if metadata.get("is_abstract"):
                 metadata_context.append("This is an abstract class/method.")
-            if metadata.get('export_type') and metadata['export_type'] != 'none':
+            if metadata.get("export_type") and metadata["export_type"] != "none":
                 metadata_context.append(f"Export type: {metadata['export_type']}")
 
         if metadata_context:
@@ -193,17 +194,17 @@ class MultiLanguageContextPrompt(ContextPromptBuilder):
 
 
 class DjangoCodeContextPrompt(ContextPromptBuilder):
-    def build_prompt(self, chunk: Dict[str, Any], symbol_index: Optional[SymbolIndex] = None) -> str:
+    def build_prompt(self, chunk: dict[str, Any], symbol_index: SymbolIndex | None = None) -> str:
         deps_info = []
         if symbol_index:
-            for dep in chunk.get('dependencies', []):
+            for dep in chunk.get("dependencies", []):
                 info = symbol_index.get_info(dep)
                 if info != "No additional info":
                     deps_info.append(f"- {dep}: {info}")
 
         deps_context = "\n".join(deps_info) if deps_info else "No additional dependency info."
 
-        context = chunk.get('context', {})
+        context = chunk.get("context", {})
         domain_ctx = context.get("domain_context", "General code")
         module_ctx = context.get("module_context", "Unknown module")
 
@@ -213,16 +214,16 @@ class DjangoCodeContextPrompt(ContextPromptBuilder):
 
         Module: {module_ctx}
         Domain context: {domain_ctx}
-        File: {chunk.get('file_path', 'unknown')}
-        Type: {chunk.get('type', 'unknown')}
-        Name: {chunk.get('name', 'unknown')}
-        Dependencies: {', '.join(chunk.get('dependencies', [])) or 'None'}
+        File: {chunk.get("file_path", "unknown")}
+        Type: {chunk.get("type", "unknown")}
+        Name: {chunk.get("name", "unknown")}
+        Dependencies: {", ".join(chunk.get("dependencies", [])) or "None"}
 
         {deps_context}
 
         Code:
-        ```{chunk.get('language', 'unknown')}
-        {chunk.get('code', '').strip()}```
+        ```{chunk.get("language", "unknown")}
+        {chunk.get("code", "").strip()}```
         Summary (one sentence, no markdown, no prefix):
         """).strip()
         return prompt
