@@ -3,15 +3,13 @@ Code chunk data structures, models, and preprocessing pipeline
 Enhanced with advanced features from mature implementations
 """
 
-from dataclasses import dataclass, asdict
-from typing import List, Dict, Any, Optional
-from pathlib import Path
 import hashlib
-import sys
+from dataclasses import asdict
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any
+
 from rich.console import Console
-from rich.panel import Panel
-from rich.progress import Progress, SpinnerColumn, TimeElapsedColumn
-from rich.table import Table
 from tqdm import tqdm
 
 console = Console()
@@ -32,24 +30,24 @@ class CodeChunk:
 
     # Basic attributes (with defaults)
     id: str = ""
-    qualified_name: Optional[str] = None
-    docstring: Optional[str] = None
-    signature: Optional[str] = None
+    qualified_name: str | None = None
+    docstring: str | None = None
+    signature: str | None = None
     complexity: int = 0
-    parent: Optional[str] = None
+    parent: str | None = None
 
     # Dependencies and references
-    dependencies: List[str] = None  # External dependencies
-    references: List[str] = None    # Symbols referenced in this chunk (NEW from enhanced.py)
-    defines: List[str] = None       # Symbols defined in this chunk (NEW from enhanced.py)
+    dependencies: list[str] = None  # External dependencies
+    references: list[str] = None  # Symbols referenced in this chunk (NEW from enhanced.py)
+    defines: list[str] = None  # Symbols defined in this chunk (NEW from enhanced.py)
 
     # Comprehensive metadata structures (NEW from enhanced.py)
-    location: Optional[Dict[str, Any]] = None        # Detailed location info (start/end line, column)
-    metadata: Optional[Dict[str, Any]] = None        # Code-specific metadata (decorators, access_modifier, etc.)
-    documentation: Optional[Dict[str, Any]] = None   # Documentation and docstring info
-    analysis: Optional[Dict[str, Any]] = None        # Code analysis (complexity, tokens, hash, etc.)
-    relationships: Optional[Dict[str, Any]] = None   # Relationship info (imports, children, etc.)
-    context: Optional[Dict[str, Any]] = None         # Context metadata (module, project, domain, hierarchy)
+    location: dict[str, Any] | None = None  # Detailed location info (start/end line, column)
+    metadata: dict[str, Any] | None = None  # Code-specific metadata (decorators, access_modifier, etc.)
+    documentation: dict[str, Any] | None = None  # Documentation and docstring info
+    analysis: dict[str, Any] | None = None  # Code analysis (complexity, tokens, hash, etc.)
+    relationships: dict[str, Any] | None = None  # Relationship info (imports, children, etc.)
+    context: dict[str, Any] | None = None  # Context metadata (module, project, domain, hierarchy)
 
     def __post_init__(self):
         if self.dependencies is None:
@@ -81,7 +79,7 @@ class CodeChunk:
     def _generate_qualified_name(self) -> str:
         """Generate fully qualified name (migrated from enhanced.py)"""
         file_stem = Path(self.file_path).stem
-        if self.type in ['class', 'function', 'method']:
+        if self.type in ["class", "function", "method"]:
             return f"{file_stem}.{self.name}"
         return self.name
 
@@ -99,30 +97,30 @@ class ChunkPreprocessor:
     def __init__(self):
         self.dedup_hashes = set()
         self.stats = {
-            'total': 0,
-            'duplicates': 0,
-            'enhanced': 0,
-            'too_large': 0,
+            "total": 0,
+            "duplicates": 0,
+            "enhanced": 0,
+            "too_large": 0,
         }
 
-    def deduplicate(self, chunks: List[Dict]) -> List[Dict]:
+    def deduplicate(self, chunks: list[dict]) -> list[dict]:
         """Remove duplicate chunks based on code content"""
         unique_chunks = []
 
         for chunk in chunks:
             # Create hash from code content
-            code_hash = hashlib.md5(chunk['code'].encode()).hexdigest()
+            code_hash = hashlib.md5(chunk["code"].encode()).hexdigest()
 
             if code_hash not in self.dedup_hashes:
                 self.dedup_hashes.add(code_hash)
                 unique_chunks.append(chunk)
             else:
-                self.stats['duplicates'] += 1
+                self.stats["duplicates"] += 1
 
         console.print(f"[yellow]Removed {self.stats['duplicates']} duplicate chunks[/yellow]")
         return unique_chunks
 
-    def enhance_chunk(self, chunk: Dict) -> Dict:
+    def enhance_chunk(self, chunk: dict) -> dict:
         """Enhance chunk with additional context for better embeddings"""
         enhanced = chunk.copy()
 
@@ -131,67 +129,67 @@ class ChunkPreprocessor:
 
         # 1. Add contextual prefix
         context_prefix = f"# {chunk['language'].upper()} {chunk['type'].upper()}"
-        if chunk.get('qualified_name'):
+        if chunk.get("qualified_name"):
             context_prefix += f": {chunk['qualified_name']}"
         parts.append(context_prefix)
 
-        if isinstance(chunk.get('context'), str):
+        if isinstance(chunk.get("context"), str):
             parts.append(f"html/css context: {chunk.get('context')}")
         else:
             # 2. Add file context
-            if chunk.get('context', {}).get('file_hierarchy'):
-                file_ctx = " > ".join(chunk['context']['file_hierarchy'])
+            if chunk.get("context", {}).get("file_hierarchy"):
+                file_ctx = " > ".join(chunk["context"]["file_hierarchy"])
                 parts.append(f"# Location: {file_ctx}")
 
             # 3. Add domain context
-            if chunk.get('context', {}).get('domain_context'):
+            if chunk.get("context", {}).get("domain_context"):
                 parts.append(f"# Purpose: {chunk['context']['domain_context']}")
 
         # 4. Add docstring if available
-        if chunk.get('docstring'):
+        if chunk.get("docstring"):
             parts.append(f'"""{chunk["docstring"]}"""')
 
         # 5. Add signature for functions
-        if chunk.get('signature'):
-            parts.append(chunk['signature'])
+        if chunk.get("signature"):
+            parts.append(chunk["signature"])
 
         # 6. Add decorators/metadata
-        if chunk.get('metadata', {}).get('decorators'):
-            parts.extend(chunk['metadata']['decorators'])
+        if chunk.get("metadata", {}).get("decorators"):
+            parts.extend(chunk["metadata"]["decorators"])
 
         # 7. Add the actual code
-        parts.append(chunk['code'])
+        parts.append(chunk["code"])
 
         # 8. Add dependencies as comments
-        if chunk.get('dependencies'):
-            deps = ", ".join(chunk['dependencies'][:5])  # Limit to 5
+        if chunk.get("dependencies"):
+            deps = ", ".join(chunk["dependencies"][:5])  # Limit to 5
             parts.append(f"# Dependencies: {deps}")
 
         # 9. Add defined symbols
-        if chunk.get('defines'):
+        if chunk.get("defines"):
             parts.append(f"# Defines: {', '.join(chunk['defines'])}")
 
         # Combine all parts
-        enhanced['embedding_text'] = "\n".join(parts)
-        enhanced['embedding_text_length'] = len(enhanced['embedding_text'])
+        enhanced["embedding_text"] = "\n".join(parts)
+        enhanced["embedding_text_length"] = len(enhanced["embedding_text"])
 
-        self.stats['enhanced'] += 1
+        self.stats["enhanced"] += 1
         return enhanced
 
-    def validate_chunk(self, chunk: Dict, max_tokens: int = 8192) -> bool:
+    def validate_chunk(self, chunk: dict, max_tokens: int = 8192) -> bool:
         """Validate chunk is suitable for embedding"""
         # Rough token estimate (1 token ≈ 4 chars)
-        estimated_tokens = len(chunk.get('embedding_text', chunk['code'])) // 4
+        estimated_tokens = len(chunk.get("embedding_text", chunk["code"])) // 4
 
         if estimated_tokens > max_tokens:
-            self.stats['too_large'] += 1
+            self.stats["too_large"] += 1
             return False
 
         return True
 
-    def process(self, chunks: List[Dict]) -> List[Dict]:
+    def process(self, chunks: list[dict]) -> list[dict]:
         """Full preprocessing pipeline"""
-        self.stats['total'] = len(chunks)
+        self.stats["total"] = len(chunks)
         console.print(f"[cyan]Starting preprocessing of {len(chunks)} chunks...[/cyan]")
 
         # 1. Deduplicate
@@ -218,30 +216,30 @@ class ChunkPreprocessor_2:
     def __init__(self):
         self.dedup_hashes = set()
         self.stats = {
-            'total': 0,
-            'duplicates': 0,
-            'enhanced': 0,
-            'too_large': 0,
+            "total": 0,
+            "duplicates": 0,
+            "enhanced": 0,
+            "too_large": 0,
         }
 
-    def deduplicate(self, chunks: List[Dict]) -> List[Dict]:
+    def deduplicate(self, chunks: list[dict]) -> list[dict]:
         """Remove duplicate chunks based on code content"""
         unique_chunks = []
 
         for chunk in chunks:
             # Create hash from code content
-            code_hash = hashlib.md5(chunk['code'].encode()).hexdigest()
+            code_hash = hashlib.md5(chunk["code"].encode()).hexdigest()
 
             if code_hash not in self.dedup_hashes:
                 self.dedup_hashes.add(code_hash)
                 unique_chunks.append(chunk)
             else:
-                self.stats['duplicates'] += 1
+                self.stats["duplicates"] += 1
 
         console.print(f"[yellow]Removed {self.stats['duplicates']} duplicate chunks[/yellow]")
         return unique_chunks
 
-    def enhance_chunk(self, chunk: Dict) -> Dict:
+    def enhance_chunk(self, chunk: dict) -> dict:
         """Enhance chunk with additional context for better embeddings"""
         enhanced = chunk.copy()
 
@@ -250,66 +248,66 @@ class ChunkPreprocessor_2:
 
         # 1. Add contextual prefix
         context_prefix = f"# {chunk['language'].upper()} {chunk['type'].upper()}"
-        if chunk.get('qualified_name'):
+        if chunk.get("qualified_name"):
             context_prefix += f": {chunk['qualified_name']}"
         parts.append(context_prefix)
-        if isinstance(chunk.get('context'), str):
+        if isinstance(chunk.get("context"), str):
             parts.append(f"html/css context: {chunk.get('context')}")
         else:
             # 2. Add file context
-            if chunk.get('context', {}).get('file_hierarchy'):
-                file_ctx = " > ".join(chunk['context']['file_hierarchy'])
+            if chunk.get("context", {}).get("file_hierarchy"):
+                file_ctx = " > ".join(chunk["context"]["file_hierarchy"])
                 parts.append(f"# Location: {file_ctx}")
 
             # 3. Add domain context
-            if chunk.get('context', {}).get('domain_context'):
+            if chunk.get("context", {}).get("domain_context"):
                 parts.append(f"# Purpose: {chunk['context']['domain_context']}")
 
         # 4. Add docstring if available
-        if chunk.get('docstring'):
+        if chunk.get("docstring"):
             parts.append(f'"""{chunk["docstring"]}"""')
 
         # 5. Add signature for functions
-        if chunk.get('signature'):
-            parts.append(chunk['signature'])
+        if chunk.get("signature"):
+            parts.append(chunk["signature"])
 
         # 6. Add decorators/metadata
-        if chunk.get('metadata', {}).get('decorators'):
-            parts.extend(chunk['metadata']['decorators'])
+        if chunk.get("metadata", {}).get("decorators"):
+            parts.extend(chunk["metadata"]["decorators"])
 
         # 7. Add the actual code
-        parts.append(chunk['code'])
+        parts.append(chunk["code"])
 
         # 8. Add dependencies as comments
-        if chunk.get('dependencies'):
-            deps = ", ".join(chunk['dependencies'][:5])  # Limit to 5
+        if chunk.get("dependencies"):
+            deps = ", ".join(chunk["dependencies"][:5])  # Limit to 5
             parts.append(f"# Dependencies: {deps}")
 
         # 9. Add defined symbols
-        if chunk.get('defines'):
+        if chunk.get("defines"):
             parts.append(f"# Defines: {', '.join(chunk['defines'])}")
 
         # Combine all parts
-        enhanced['embedding_text'] = "\n".join(parts)
-        enhanced['embedding_text_length'] = len(enhanced['embedding_text'])
+        enhanced["embedding_text"] = "\n".join(parts)
+        enhanced["embedding_text_length"] = len(enhanced["embedding_text"])
 
-        self.stats['enhanced'] += 1
+        self.stats["enhanced"] += 1
         return enhanced
 
-    def validate_chunk(self, chunk: Dict, max_tokens: int = 8192) -> bool:
+    def validate_chunk(self, chunk: dict, max_tokens: int = 8192) -> bool:
         """Validate chunk is suitable for embedding"""
         # Rough token estimate (1 token ≈ 4 chars)
-        estimated_tokens = len(chunk.get('embedding_text', chunk['code'])) // 4
+        estimated_tokens = len(chunk.get("embedding_text", chunk["code"])) // 4
 
         if estimated_tokens > max_tokens:
-            self.stats['too_large'] += 1
+            self.stats["too_large"] += 1
             return False
 
         return True
 
-    def process(self, chunks: List[Dict]) -> List[Dict]:
+    def process(self, chunks: list[dict]) -> list[dict]:
         """Full preprocessing pipeline"""
-        self.stats['total'] = len(chunks)
+        self.stats["total"] = len(chunks)
         console.print(f"[cyan]Starting preprocessing of {len(chunks)} chunks...[/cyan]")
 
         # 1. Deduplicate
