@@ -2,24 +2,23 @@
 Metadata extraction from code chunks
 """
 
-from typing import Optional
 from .chunk import CodeChunk
 
 
 class MetadataExtractor:
     """Handles extraction of semantic metadata from code"""
 
-    def extract_docstring(self, node, code_bytes: bytes) -> Optional[str]:
+    def extract_docstring(self, node, code_bytes: bytes) -> str | None:
         """Extract docstring from AST node"""
-        if node.type in ['function_definition', 'class_definition']:
+        if node.type in ["function_definition", "class_definition"]:
             # This would be language-specific
             # For now, return None as it's handled in analyzer
             pass
         return None
 
-    def extract_signature(self, code: str) -> Optional[str]:
+    def extract_signature(self, code: str) -> str | None:
         """Extract function/method signature"""
-        lines = code.strip().split('\n')
+        lines = code.strip().split("\n")
         if lines:
             return lines[0].strip()
         return None
@@ -27,7 +26,7 @@ class MetadataExtractor:
     def extract_complexity(self, code: str) -> int:
         """Calculate cyclomatic complexity"""
         complexity = 1
-        keywords = ['if ', 'elif ', 'else ', 'for ', 'while ', 'case ', 'catch ', ' && ', ' || ']
+        keywords = ["if ", "elif ", "else ", "for ", "while ", "case ", "catch ", " && ", " || "]
         for keyword in keywords:
             complexity += code.count(keyword)
         return complexity
@@ -68,7 +67,7 @@ class MetadataExtractor:
     def enhance_chunk_metadata(self, chunk: CodeChunk) -> CodeChunk:
         """Add additional metadata to a chunk"""
         # Add tags
-        if not hasattr(chunk, 'tags') or not chunk.tags:
+        if not hasattr(chunk, "tags") or not chunk.tags:
             chunk.tags = self.extract_tags_and_categories(chunk)
 
         # Ensure all required fields are present
@@ -78,11 +77,11 @@ class MetadataExtractor:
         if chunk.signature is None:
             chunk.signature = self.extract_signature(chunk.code)
 
-        if not hasattr(chunk, 'metadata'):
+        if not hasattr(chunk, "metadata"):
             chunk.metadata = {
-                'line_count': chunk.end_line - chunk.start_line + 1,
-                'has_tests': any('test' in chunk.name.lower() for chunk in [chunk]),  # simplistic check
-                'is_entry_point': chunk.name in ['main', '__main__', 'app', 'application']
+                "line_count": chunk.end_line - chunk.start_line + 1,
+                "has_tests": any("test" in chunk.name.lower() for chunk in [chunk]),  # simplistic check
+                "is_entry_point": chunk.name in ["main", "__main__", "app", "application"],
             }
 
         return chunk
