@@ -1,33 +1,38 @@
-
 """
 Hybrid Search Implementation for Qdrant
 Combines Dense Vectors (semantic) + Sparse Vectors (BM25 keyword)
 Migrated from old_code
 """
 
-from typing import List, Dict, Any, Optional, Tuple
-from dataclasses import dataclass
+import pathlib
 import re
-from qdrant_client import QdrantClient
-from qdrant_client.models import (
-    Distance, VectorParams, PointStruct,
-    Filter, FieldCondition, MatchValue, MatchAny,
-    SparseVector, SparseVectorParams, SparseIndexParams,
-    NamedVector, NamedSparseVector,
-    Modifier, SearchRequest, QueryRequest,
-    FusionQuery, Prefetch, Query
-)
-import sys
-from rich.console import Console
-from rich.table import Table
-from rich.panel import Panel
-from rich.syntax import Syntax
-from rich.progress import Progress, SpinnerColumn, TextColumn, BarColumn, TaskProgressColumn
-
-import requests
-import numpy as np
 import time
-from ..embedding.embedder import EmbeddingGenerator, EmbeddingConfig
+from dataclasses import dataclass
+
+import numpy as np
+from qdrant_client import QdrantClient
+from qdrant_client.models import Distance
+from qdrant_client.models import FieldCondition
+from qdrant_client.models import Filter
+from qdrant_client.models import FusionQuery
+from qdrant_client.models import MatchValue
+from qdrant_client.models import PointStruct
+from qdrant_client.models import Prefetch
+from qdrant_client.models import SparseIndexParams
+from qdrant_client.models import SparseVector
+from qdrant_client.models import SparseVectorParams
+from qdrant_client.models import VectorParams
+from rich.console import Console
+from rich.panel import Panel
+from rich.progress import BarColumn
+from rich.progress import Progress
+from rich.progress import SpinnerColumn
+from rich.progress import TaskProgressColumn
+from rich.progress import TextColumn
+from rich.table import Table
+
+from ..embedding.embedder import EmbeddingConfig
+from ..embedding.embedder import EmbeddingGenerator
 
 console = Console()
 
@@ -35,6 +40,7 @@ console = Console()
 @dataclass
 class HybridSearchConfig:
     """Configuration for hybrid search"""
+
     # Dense vector (semantic) weight
     dense_weight: float = 0.7
 
@@ -53,6 +59,7 @@ class HybridSearchConfig:
     # Score threshold
     min_score: float = 0.5
 
+
 class BM25SparseEncoder:
     """
     Simple BM25 sparse vector encoder
@@ -66,11 +73,11 @@ class BM25SparseEncoder:
         self.idf = {}
         self.avg_doc_len = 0
 
-    def tokenize(self, text: str) -> List[str]:
+    def tokenize(self, text: str) -> list[str]:
         """Simple tokenization"""
         # Convert to lowercase, split on non-alphanumeric
         text = text.lower()
-        tokens = re.findall(r'\b\w+\b', text)
+        tokens = re.findall(r"\b\w+\b", text)
         return tokens
 
     def build_vocab_from_collection(self, client: QdrantClient, collection_name: str):
@@ -88,27 +95,19 @@ class BM25SparseEncoder:
         term_doc_freq = {}
 
         with Progress(
-              SpinnerColumn(),
-              TextColumn("[progress.description]{task.description}"),
-              BarColumn(),
-              TaskProgressColumn(),
-              TextColumn("[progress.percentage]{task.percentage:>3.0f}%"),
-              console=console
-            ) as progress:
-
+            SpinnerColumn(),
+            TextColumn("[progress.description]{task.description}"),
+            BarColumn(),
+            TaskProgressColumn(),
+            TextColumn("[progress.percentage]{task.percentage:>3.0f}%"),
+            console=console,
+        ) as progress:
             # Main document processing task
-            doc_task = progress.add_task(
-                f"[red]Processing documents...[/red]\n",
-                total=total_points
-            )
+            doc_task = progress.add_task(f"[red]Processing documents...[/red]\n", total=total_points)
 
             while True:
                 result = client.scroll(
-                    collection_name=collection_name,
-                    limit=100,
-                    offset=offset,
-                    with_payload=True,
-                    with_vectors=False
+                    collection_name=collection_name, limit=100, offset=offset, with_payload=True, with_vectors=False
                 )
 
                 points, next_offset = result
@@ -120,15 +119,14 @@ class BM25SparseEncoder:
                     break
 
                 for point in points:
-
                     # Combine relevant text fields
                     text_parts = [
-                        point.payload.get('code', ''),
-                        point.payload.get('docstring', ''),
-                        point.payload.get('qualified_name', ''),
-                        point.payload.get('signature', '')
+                        point.payload.get("code", ""),
+                        point.payload.get("docstring", ""),
+                        point.payload.get("qualified_name", ""),
+                        point.payload.get("signature", ""),
                     ]
-                    text = ' '.join(filter(None, text_parts))
+                    text = " ".join(filter(None, text_parts))
 
                     tokens = self.tokenize(text)
                     doc_count += 1
@@ -148,8 +146,7 @@ class BM25SparseEncoder:
             print("here")
             if term_doc_freq:
                 idf_task = progress.add_task(
-                    f"[green]Calculating IDF for {len(term_doc_freq)} terms...",
-                    total=len(term_doc_freq)
+                    f"[green]Calculating IDF for {len(term_doc_freq)} terms...", total=len(term_doc_freq)
                 )
                 # Calculate IDF
                 self.avg_doc_len = total_length / doc_count if doc_count > 0 else 0
@@ -187,12 +184,9 @@ class BM25SparseEncoder:
                 indices.append(self.vocab[term])
                 values.append(float(score))
 
-        return SparseVector(
-            indices=indices,
-            values=values
-        )
+        return SparseVector(indices=indices, values=values)
 
-    def build_vocab_from_texts(self, texts: List[str]):
+    def build_vocab_from_texts(self, texts: list[str]):
         """Build vocabulary and IDF from raw texts (not from Qdrant)"""
         console.print(f"[cyan]Building BM25 vocabulary from {len(texts)} documents...[/cyan]")
 
@@ -206,7 +200,7 @@ class BM25SparseEncoder:
             BarColumn(),
             TaskProgressColumn(),
             TextColumn("[progress.percentage]{task.percentage:>3.0f}%"),
-            console=console
+            console=console,
         ) as progress:
             doc_task = progress.add_task("[red]Processing documents...[/red]", total=doc_count)
 
@@ -220,8 +214,7 @@ class BM25SparseEncoder:
 
             if term_doc_freq:
                 idf_task = progress.add_task(
-                    f"[green]Calculating IDF for {len(term_doc_freq)} terms...",
-                    total=len(term_doc_freq)
+                    f"[green]Calculating IDF for {len(term_doc_freq)} terms...", total=len(term_doc_freq)
                 )
                 self.avg_doc_len = total_length / doc_count if doc_count > 0 else 0
 
@@ -232,24 +225,17 @@ class BM25SparseEncoder:
 
         console.print(f"[green]✓ Vocabulary built: {len(self.vocab)} terms from {doc_count} documents[/green]")
 
+
 class HybridSearchEngine:
     """
     Advanced hybrid search combining dense and sparse vectors
     """
 
-    def __init__(
-        self,
-        qdrant_client: QdrantClient,
-        embedding_generator,
-        config: HybridSearchConfig = None
-    ):
+    def __init__(self, qdrant_client: QdrantClient, embedding_generator, config: HybridSearchConfig = None):
         self.client = qdrant_client
         self.embedding_gen = embedding_generator
         self.config = config or HybridSearchConfig()
-        self.bm25_encoder = BM25SparseEncoder(
-            k1=self.config.bm25_k1,
-            b=self.config.bm25_b
-        )
+        self.bm25_encoder = BM25SparseEncoder(k1=self.config.bm25_k1, b=self.config.bm25_b)
 
     def create_hybrid_collection(self, collection_name: str, dense_dim: int = 768):
         """Create a new collection with named dense + sparse vectors"""
@@ -257,18 +243,12 @@ class HybridSearchEngine:
 
         self.client.recreate_collection(
             collection_name=collection_name,
-            vectors_config={
-                "text-dense": VectorParams(size=dense_dim, distance=Distance.COSINE)
-            },
-            sparse_vectors_config={
-                "text-sparse": SparseVectorParams(
-                    index=SparseIndexParams(on_disk=False)
-                )
-            }
+            vectors_config={"text-dense": VectorParams(size=dense_dim, distance=Distance.COSINE)},
+            sparse_vectors_config={"text-sparse": SparseVectorParams(index=SparseIndexParams(on_disk=False))},
         )
         console.print(f"[green]✓ Hybrid collection created[/green]")
 
-    def reindex_with_sparse_vectors(self, collection_name: str, chunks: List[Dict]):
+    def reindex_with_sparse_vectors(self, collection_name: str, chunks: list[dict]):
         """
         Add sparse vectors to existing points
         """
@@ -278,52 +258,45 @@ class HybridSearchEngine:
 
         batch_size = 100
         for i in tqdm(range(0, len(chunks), batch_size), desc="Reindexing"):
-            batch = chunks[i:i + batch_size]
+            batch = chunks[i : i + batch_size]
             points = []
 
             for chunk in batch:
                 # Generate sparse vector
                 text_parts = [
-                    chunk.get('code', ''),
-                    chunk.get('docstring', ''),
-                    chunk.get('qualified_name', ''),
+                    chunk.get("code", ""),
+                    chunk.get("docstring", ""),
+                    chunk.get("qualified_name", ""),
                 ]
-                text = ' '.join(filter(None, text_parts))
+                text = " ".join(filter(None, text_parts))
                 sparse_vector = self.bm25_encoder.encode(text)
 
                 # Create point with both dense and sparse vectors
                 point = PointStruct(
                     id=int(chunk["id"], 16),
                     vector={
-                        "text-dense": chunk['embedding'],  # Existing dense vector
-                        "text-sparse": sparse_vector
+                        "text-dense": chunk["embedding"],  # Existing dense vector
+                        "text-sparse": sparse_vector,
                     },
-                    payload=self._prepare_payload(chunk)
+                    payload=self._prepare_payload(chunk),
                 )
                 points.append(point)
 
             # Upsert batch
-            self.client.upsert(
-                collection_name=collection_name,
-                points=points
-            )
+            self.client.upsert(collection_name=collection_name, points=points)
 
         console.print(f"[green]✓ Reindexing complete[/green]")
 
-    def _prepare_payload(self, chunk: Dict) -> Dict:
+    def _prepare_payload(self, chunk: dict) -> dict:
         """Prepare payload (remove embedding)"""
         payload = chunk.copy()
-        payload.pop('embedding', None)
-        payload.pop('embedding_text', None)
+        payload.pop("embedding", None)
+        payload.pop("embedding_text", None)
         return payload
 
     def hybrid_search(
-        self,
-        collection_name: str,
-        query_text: str,
-        filters: Optional[Filter] = None,
-        limit: int = None
-    ) -> List[Dict]:
+        self, collection_name: str, query_text: str, filters: Filter | None = None, limit: int = None
+    ) -> list[dict]:
         """
         Perform hybrid search (dense + sparse vectors)
         """
@@ -344,21 +317,11 @@ class HybridSearchEngine:
             results = self.client.query_points(
                 collection_name=collection_name,
                 prefetch=[
-                    Prefetch(
-                        query=query_embedding,
-                        using="text-dense",
-                        limit=self.config.top_k_dense,
-                        filter=filters
-                    ),
-                    Prefetch(
-                        query=query_sparse,
-                        using="text-sparse",
-                        limit=self.config.top_k_sparse,
-                        filter=filters
-                    )
+                    Prefetch(query=query_embedding, using="text-dense", limit=self.config.top_k_dense, filter=filters),
+                    Prefetch(query=query_sparse, using="text-sparse", limit=self.config.top_k_sparse, filter=filters),
                 ],
                 query=FusionQuery(fusion="rrf"),  # Reciprocal Rank Fusion
-                limit=limit
+                limit=limit,
             )
 
             return results.points
@@ -366,47 +329,33 @@ class HybridSearchEngine:
         except Exception as e:
             console.print(f"[yellow]Hybrid search not available, falling back to dense-only: {e}[/yellow]")
             # Fallback to dense-only search
-            return self._dense_search_fallback(
-                collection_name, query_embedding, filters, limit
-            )
+            return self._dense_search_fallback(collection_name, query_embedding, filters, limit)
 
     def _dense_search_fallback(
-        self,
-        collection_name: str,
-        query_vector: List[float],
-        filters: Optional[Filter],
-        limit: int
+        self, collection_name: str, query_vector: list[float], filters: Filter | None, limit: int
     ):
         """Fallback to dense-only search"""
         return self.client.search(
             collection_name=collection_name,
             query_vector=("text-dense", query_vector),
             query_filter=filters,
-            limit=limit
+            limit=limit,
         )
 
-    def compare_search_methods(
-        self,
-        collection_name: str,
-        query: str,
-        limit: int = 5
-    ):
+    def compare_search_methods(self, collection_name: str, query: str, limit: int = 5):
         """
         Compare dense-only vs hybrid search results
         """
-        console.print(Panel.fit(
-            f"[bold cyan]Comparing Search Methods[/bold cyan]\nQuery: '{query}'",
-            border_style="cyan"
-        ))
+        console.print(
+            Panel.fit(f"[bold cyan]Comparing Search Methods[/bold cyan]\nQuery: '{query}'", border_style="cyan")
+        )
 
         # 1. Dense-only search
         query_embedding = self.embedding_gen.generate_embedding(query)
 
         console.print("\n[yellow]═══ Dense-Only Search (Vector Similarity) ═══[/yellow]")
         dense_results = self.client.search(
-            collection_name=collection_name,
-            query_vector=("text-dense", query_embedding),
-            limit=limit
+            collection_name=collection_name, query_vector=("text-dense", query_embedding), limit=limit
         )
 
         self._display_results(dense_results, "Dense")
@@ -422,11 +371,7 @@ class HybridSearchEngine:
 
     def _display_results(self, results, method_name: str):
         """Display search results in a table"""
-        table = Table(
-            show_header=True,
-            header_style="bold magenta",
-            title=f"{method_name} Search Results"
-        )
+        table = Table(show_header=True, header_style="bold magenta", title=f"{method_name} Search Results")
         table.add_column("Rank", width=6)
         table.add_column("Score", width=8)
         table.add_column("Name", width=30)
@@ -437,9 +382,9 @@ class HybridSearchEngine:
             table.add_row(
                 str(i),
                 f"{result.score:.4f}",
-                result.payload.get('qualified_name', result.payload.get('name', 'N/A'))[:30],
-                result.payload.get('type', 'N/A'),
-                result.payload.get('file_path', 'N/A')[:40]
+                result.payload.get("qualified_name", result.payload.get("name", "N/A"))[:30],
+                result.payload.get("type", "N/A"),
+                result.payload.get("file_path", "N/A")[:40],
             )
 
         console.print(table)
@@ -469,42 +414,28 @@ class HybridSearchEngine:
         self,
         collection_name: str,
         query: str,
-        language: Optional[str] = None,
-        chunk_type: Optional[str] = None,
-        min_complexity: Optional[str] = None,
-        file_pattern: Optional[str] = None,
-        limit: int = 10
-    ) -> List[Dict]:
+        language: str | None = None,
+        chunk_type: str | None = None,
+        min_complexity: str | None = None,
+        file_pattern: str | None = None,
+        limit: int = 10,
+    ) -> list[dict]:
         """
         Advanced search with multiple filters
         """
         must_conditions = []
 
         if language:
-            must_conditions.append(
-                FieldCondition(key="language", match=MatchValue(value=language))
-            )
+            must_conditions.append(FieldCondition(key="language", match=MatchValue(value=language)))
 
         if chunk_type:
-            must_conditions.append(
-                FieldCondition(key="type", match=MatchValue(value=chunk_type))
-            )
+            must_conditions.append(FieldCondition(key="type", match=MatchValue(value=chunk_type)))
 
         if min_complexity:
-            must_conditions.append(
-                FieldCondition(
-                    key="complexity",
-                    range={"gte": min_complexity}
-                )
-            )
+            must_conditions.append(FieldCondition(key="complexity", range={"gte": min_complexity}))
 
         if file_pattern:
-            must_conditions.append(
-                FieldCondition(
-                    key="file_path",
-                    match=MatchValue(value=file_pattern)
-                )
-            )
+            must_conditions.append(FieldCondition(key="file_path", match=MatchValue(value=file_pattern)))
 
         filters = Filter(must=must_conditions) if must_conditions else None
 
@@ -519,27 +450,26 @@ def setup_hybrid_collection(collection_name: str, chunks_path: str):
     embedding_gen = EmbeddingGenerator(EmbeddingConfig())
 
     hybrid_engine = HybridSearchEngine(
-        qdrant_client=client,
-        embedding_generator=embedding_gen,
-        config=HybridSearchConfig(final_top_k=10)
+        qdrant_client=client, embedding_generator=embedding_gen, config=HybridSearchConfig(final_top_k=10)
     )
 
     # Load chunks
     import json
-    with open(chunks_path, 'r') as f:
+
+    with pathlib.Path(chunks_path).open("r") as f:
         data = json.load(f)
-        chunks = data.get('chunks', [])
+        chunks = data.get("chunks", [])
 
     # Build BM25 vocab from raw texts
     console.print("[cyan]Building BM25 vocabulary from chunks...[/cyan]")
     all_texts = []
     for chunk in chunks:
         text_parts = [
-            chunk.get('code', ''),
-            chunk.get('docstring', ''),
-            chunk.get('qualified_name', ''),
+            chunk.get("code", ""),
+            chunk.get("docstring", ""),
+            chunk.get("qualified_name", ""),
         ]
-        text = ' '.join(filter(None, text_parts))
+        text = " ".join(filter(None, text_parts))
         all_texts.append(text)
     hybrid_engine.bm25_encoder.build_vocab_from_texts(all_texts)
 
@@ -549,11 +479,13 @@ def setup_hybrid_collection(collection_name: str, chunks_path: str):
     # Reindex all chunks
     hybrid_engine.reindex_with_sparse_vectors(collection_name, chunks)
 
-    console.print(Panel.fit(
-        "[bold green]✓ Hybrid Search Enabled![/bold green]\n"
-        "Your collection now supports both semantic and keyword search.",
-        border_style="green"
-    ))
+    console.print(
+        Panel.fit(
+            "[bold green]✓ Hybrid Search Enabled![/bold green]\n"
+            "Your collection now supports both semantic and keyword search.",
+            border_style="green",
+        )
+    )
 
 
 def main():
@@ -561,31 +493,30 @@ def main():
     embedding_gen = EmbeddingGenerator(EmbeddingConfig())
 
     hybrid_engine = HybridSearchEngine(
-        qdrant_client=client,
-        embedding_generator=embedding_gen,
-        config=HybridSearchConfig(final_top_k=10)
+        qdrant_client=client, embedding_generator=embedding_gen, config=HybridSearchConfig(final_top_k=10)
     )
 
     collection_name = input("\nEnter new collection name for hybrid search: ").strip()
 
     # Load chunks
     import json
+
     # chunks_file = input("Enter path to embedded chunks JSON: ").strip()
     chunks_file = "parsed_chunks_tipsy_2_embedded.json"
-    with open(chunks_file, 'r') as f:
+    with pathlib.Path(chunks_file).open("r") as f:
         data = json.load(f)
-        chunks = data.get('chunks', [])
+        chunks = data.get("chunks", [])
 
     # Build BM25 vocab from raw texts
     console.print("[cyan]Building BM25 vocabulary from chunks...[/cyan]")
     all_texts = []
     for chunk in chunks:
         text_parts = [
-            chunk.get('code', ''),
-            chunk.get('docstring', ''),
-            chunk.get('qualified_name', ''),
+            chunk.get("code", ""),
+            chunk.get("docstring", ""),
+            chunk.get("qualified_name", ""),
         ]
-        text = ' '.join(filter(None, text_parts))
+        text = " ".join(filter(None, text_parts))
         all_texts.append(text)
     hybrid_engine.bm25_encoder.build_vocab_from_texts(all_texts)
 
@@ -600,11 +531,13 @@ def main():
     if test_query:
         hybrid_engine.compare_search_methods(collection_name, test_query, limit=5)
 
-    console.print(Panel.fit(
-        "[bold green]✓ Hybrid Search Enabled![/bold green]\n"
-        "Your collection now supports both semantic and keyword search.",
-        border_style="green"
-    ))
+    console.print(
+        Panel.fit(
+            "[bold green]✓ Hybrid Search Enabled![/bold green]\n"
+            "Your collection now supports both semantic and keyword search.",
+            border_style="green",
+        )
+    )
 
 
 if __name__ == "__main__":
