@@ -4,21 +4,25 @@ Qdrant vector database integration for indexing and searching - Enhanced version
 
 import json
 import time
-from typing import List, Dict, Any, Optional
 from pathlib import Path
-from tqdm import tqdm
-from rich.console import Console
-from rich.table import Table
-from rich.panel import Panel
+
 from qdrant_client import QdrantClient
-from qdrant_client.models import (
-    Distance, VectorParams, PointStruct,
-    Filter, FieldCondition, MatchValue,
-    PayloadSchemaType, TextIndexParams, TextIndexType,
-    CreateCollection, SparseVectorParams, SparseIndexParams
-)
+from qdrant_client.models import FieldCondition
+from qdrant_client.models import Filter
+from qdrant_client.models import MatchValue
+from qdrant_client.models import PayloadSchemaType
+from qdrant_client.models import PointStruct
+from qdrant_client.models import SparseIndexParams
+from qdrant_client.models import SparseVectorParams
+from qdrant_client.models import VectorParams
+from rich.console import Console
+from rich.panel import Panel
+from rich.table import Table
+from tqdm import tqdm
+
 # Import consolidated configuration for backward compatibility
-from ..config import EmbeddingConfig, QdrantConfig
+from ..config import EmbeddingConfig
+from ..config import QdrantConfig
 
 # Import embedding generator from correct location
 from ..embedding.embedder import EmbeddingGenerator
@@ -34,10 +38,10 @@ class QdrantIndexer:
         self.client = QdrantClient(host=config.host, port=config.port)
         self.collections = {}
         self.stats = {
-            'collections_created': 0,
-            'chunks_indexed': 0,
-            'indexing_errors': 0,
-            'searches_performed': 0,
+            "collections_created": 0,
+            "chunks_indexed": 0,
+            "indexing_errors": 0,
+            "searches_performed": 0,
         }
 
     def health_check(self) -> bool:
@@ -56,15 +60,15 @@ class QdrantIndexer:
         collection_configs = {
             f"{self.config.collection_prefix}_functions": {
                 "description": "Individual functions and methods",
-                "optimization": "High-frequency function lookups"
+                "optimization": "High-frequency function lookups",
             },
             f"{self.config.collection_prefix}_classes": {
                 "description": "Class definitions and structure",
-                "optimization": "Class hierarchy and inheritance"
+                "optimization": "Class hierarchy and inheritance",
             },
             f"{self.config.collection_prefix}_modules": {
                 "description": "Module and file-level code",
-                "optimization": "File-level context and imports"
+                "optimization": "File-level context and imports",
             },
         }
 
@@ -78,7 +82,9 @@ class QdrantIndexer:
 
                 # Update existing collection if needed
                 if collection_info.config.params.vectors.size != embedding_dim:
-                    console.print(f"[yellow]Updating vector dimension from {collection_info.config.params.vectors.size} to {embedding_dim}[/yellow]")
+                    console.print(
+                        f"[yellow]Updating vector dimension from {collection_info.config.params.vectors.size} to {embedding_dim}[/yellow]"
+                    )
 
             except Exception:
                 # Create new collection
@@ -92,17 +98,11 @@ class QdrantIndexer:
                         ),
                         sparse_vectors_config={
                             # Sparse identifiers (variable/function/class names)
-                            "symbols": SparseVectorParams(
-                                index=SparseIndexParams(on_disk=False)
-                            ),
+                            "symbols": SparseVectorParams(index=SparseIndexParams(on_disk=False)),
                             # BM25-like text search (docstrings, comments)
-                            "bm25": SparseVectorParams(
-                                index=SparseIndexParams(on_disk=False)
-                            ),
-                            "text-sparse": SparseVectorParams(
-                                index=SparseIndexParams(on_disk=False)
-                            ),
-                        }
+                            "bm25": SparseVectorParams(index=SparseIndexParams(on_disk=False)),
+                            "text-sparse": SparseVectorParams(index=SparseIndexParams(on_disk=False)),
+                        },
                     )
 
                     # Create payload indexes for filtering
@@ -111,17 +111,15 @@ class QdrantIndexer:
                         self.client.create_payload_index(
                             collection_name=collection_name,
                             field_name="language",
-                            field_schema=PayloadSchemaType.KEYWORD
+                            field_schema=PayloadSchemaType.KEYWORD,
                         )
                         self.client.create_payload_index(
-                            collection_name=collection_name,
-                            field_name="type",
-                            field_schema=PayloadSchemaType.KEYWORD
+                            collection_name=collection_name, field_name="type", field_schema=PayloadSchemaType.KEYWORD
                         )
                         self.client.create_payload_index(
                             collection_name=collection_name,
                             field_name="file_path",
-                            field_schema=PayloadSchemaType.KEYWORD
+                            field_schema=PayloadSchemaType.KEYWORD,
                         )
                         self.client.create_payload_index(
                             collection_name=collection_name,
@@ -129,7 +127,7 @@ class QdrantIndexer:
                             field_schema=PayloadSchemaType.TEXT,
                         )
 
-                    self.stats['collections_created'] += 1
+                    self.stats["collections_created"] += 1
                     console.print(f"[green]✓ Created collection: {collection_name}[/green]")
                     console.print(f"  {config['description']}")
                     console.print(f"  Optimization: {config['optimization']}")
@@ -140,51 +138,50 @@ class QdrantIndexer:
 
             self.collections[collection_name] = config
 
-    def _get_collection_for_chunk(self, chunk: Dict) -> str:
+    def _get_collection_for_chunk(self, chunk: dict) -> str:
         """Determine which collection a chunk should go into with enhanced logic"""
-        chunk_type = chunk.get('type', 'unknown')
+        chunk_type = chunk.get("type", "unknown")
 
         # More specific type handling
-        if chunk_type in ['function', 'method', 'lambda']:
+        if chunk_type in ["function", "method", "lambda"]:
             return f"{self.config.collection_prefix}_functions"
-        elif chunk_type == 'class':
+        if chunk_type == "class":
             return f"{self.config.collection_prefix}_classes"
-        elif chunk_type in ['module', 'file', 'script']:
+        if chunk_type in ["module", "file", "script"]:
             return f"{self.config.collection_prefix}_modules"
-        else:
-            # Default to modules for unknown types
-            console.print(f"[yellow]Unknown chunk type '{chunk_type}', defaulting to modules collection[/yellow]")
-            return f"{self.config.collection_prefix}_modules"
+        # Default to modules for unknown types
+        console.print(f"[yellow]Unknown chunk type '{chunk_type}', defaulting to modules collection[/yellow]")
+        return f"{self.config.collection_prefix}_modules"
 
-    def _prepare_payload(self, chunk: Dict) -> Dict:
+    def _prepare_payload(self, chunk: dict) -> dict:
         """Prepare chunk payload for Qdrant with enhanced metadata handling"""
         payload = chunk.copy()
 
         # Remove embedding (stored separately)
-        payload.pop('embedding', None)
-        payload.pop('embedding_text', None)  # Don't need full text in payload
+        payload.pop("embedding", None)
+        payload.pop("embedding_text", None)  # Don't need full text in payload
 
         # Keep essential fields easily filterable with defaults
         essential = {
-            'id': chunk['id'],
-            'name': chunk['name'],
-            'type': chunk['type'],
-            'language': chunk['language'],
-            'file_path': chunk['file_path'],
-            'qualified_name': chunk.get('qualified_name', ''),
-            'code': chunk['code'],  # Keep original code
-            'start_line': chunk['start_line'],
-            'end_line': chunk['end_line'],
-            'complexity': chunk.get('complexity', 0),
-            'docstring': chunk.get('docstring'),
-            'signature': chunk.get('signature'),
-            'dependencies': chunk.get('dependencies', []),
-            'context': chunk.get('context', {}),
-            'metadata': chunk.get('metadata', {}),
-            'relationships': chunk.get('relationships', {}),
-            'analysis': chunk.get('analysis', {}),
-            'embedding_quality': chunk.get('embedding_quality', 'unknown'),
-            'indexed_at': chunk.get('embedding_timestamp', time.time()),
+            "id": chunk["id"],
+            "name": chunk["name"],
+            "type": chunk["type"],
+            "language": chunk["language"],
+            "file_path": chunk["file_path"],
+            "qualified_name": chunk.get("qualified_name", ""),
+            "code": chunk["code"],  # Keep original code
+            "start_line": chunk["start_line"],
+            "end_line": chunk["end_line"],
+            "complexity": chunk.get("complexity", 0),
+            "docstring": chunk.get("docstring"),
+            "signature": chunk.get("signature"),
+            "dependencies": chunk.get("dependencies", []),
+            "context": chunk.get("context", {}),
+            "metadata": chunk.get("metadata", {}),
+            "relationships": chunk.get("relationships", {}),
+            "analysis": chunk.get("analysis", {}),
+            "embedding_quality": chunk.get("embedding_quality", "unknown"),
+            "indexed_at": chunk.get("embedding_timestamp", time.time()),
         }
 
         return essential
@@ -198,14 +195,13 @@ class QdrantIndexer:
             # Adaptive batch sizing based on collection size
             if current_size < 1000:
                 return 50  # Small collections, smaller batches
-            elif current_size < 10000:
+            if current_size < 10000:
                 return 100  # Medium collections
-            else:
-                return 200  # Large collections, larger batches
+            return 200  # Large collections, larger batches
         except:
             return 100  # Default batch size
 
-    def index_chunks(self, chunks: List[Dict], batch_size: Optional[int] = None):
+    def index_chunks(self, chunks: list[dict], batch_size: int | None = None):
         """Index chunks in Qdrant with enhanced error handling and progress tracking"""
         console.print(f"[cyan]Enhanced indexing of {len(chunks)} chunks in Qdrant...[/cyan]")
 
@@ -237,38 +233,37 @@ class QdrantIndexer:
             # Prepare points
             points = []
             for chunk in coll_chunks:
-                if 'embedding' not in chunk:
+                if "embedding" not in chunk:
                     console.print(f"[yellow]Skipping chunk without embedding: {chunk.get('name', 'unknown')}[/yellow]")
                     continue
 
                 point = PointStruct(
                     id=int(chunk["id"], 16),  # Use chunk ID
-                    vector=chunk['embedding'],
-                    payload=self._prepare_payload(chunk)
+                    vector=chunk["embedding"],
+                    payload=self._prepare_payload(chunk),
                 )
                 points.append(point)
 
             # Batch upload with enhanced error handling
             batches_uploaded = 0
             for i in tqdm(range(0, len(points), batch_size), desc=f"Uploading to {collection_name}"):
-                batch = points[i:i + batch_size]
+                batch = points[i : i + batch_size]
                 try:
-                    self.client.upsert(
-                        collection_name=collection_name,
-                        points=batch
-                    )
+                    self.client.upsert(collection_name=collection_name, points=batch)
                     total_indexed += len(batch)
                     batches_uploaded += 1
 
                 except Exception as e:
-                    self.stats['indexing_errors'] += 1
+                    self.stats["indexing_errors"] += 1
                     console.print(f"[red]Failed to upload batch to {collection_name}: {e}[/red]")
                     # Try smaller batch size
                     if batch_size > 10:
                         console.print(f"[yellow]Retrying with smaller batch size...[/yellow]")
                         batch_size = max(10, batch_size // 2)
 
-            console.print(f"[green]✓ Completed {collection_name}: {len(coll_chunks)} chunks in {batches_uploaded} batches[/green]")
+            console.print(
+                f"[green]✓ Completed {collection_name}: {len(coll_chunks)} chunks in {batches_uploaded} batches[/green]"
+            )
 
         # Enhanced statistics
         total_time = time.time() - start_time
@@ -281,18 +276,21 @@ class QdrantIndexer:
         console.print(f"  Rate: {avg_time_per_chunk:.3f}s per chunk")
         console.print(f"  Errors: {self.stats['indexing_errors']}")
 
-        self.stats['chunks_indexed'] = total_indexed
+        self.stats["chunks_indexed"] = total_indexed
 
-    def advanced_search(self, query: str, limit: int = 10, filters: Optional[Dict] = None,
-                       collection_filter: Optional[str] = None) -> Dict[str, List]:
+    def advanced_search(
+        self, query: str, limit: int = 10, filters: dict | None = None, collection_filter: str | None = None
+    ) -> dict[str, list]:
         """Perform advanced search with filtering and ranking"""
-        console.print(Panel.fit(
-            f"[bold cyan]Advanced Search: '{query}'[/bold cyan]",
-            subtitle=f"Limit: {limit} | Filters: {filters} | Collection: {collection_filter or 'all'}",
-            border_style="cyan"
-        ))
+        console.print(
+            Panel.fit(
+                f"[bold cyan]Advanced Search: '{query}'[/bold cyan]",
+                subtitle=f"Limit: {limit} | Filters: {filters} | Collection: {collection_filter or 'all'}",
+                border_style="cyan",
+            )
+        )
 
-        self.stats['searches_performed'] += 1
+        self.stats["searches_performed"] += 1
 
         # Generate query embedding
         embedding_gen = EmbeddingGenerator(EmbeddingConfig())
@@ -308,15 +306,14 @@ class QdrantIndexer:
             conditions = []
             for field, value in filters.items():
                 if isinstance(value, list):
-                    conditions.append(FieldCondition(
-                        key=field,
-                        match=MatchValue(value=value[0])  # Simple match, could be extended
-                    ))
+                    conditions.append(
+                        FieldCondition(
+                            key=field,
+                            match=MatchValue(value=value[0]),  # Simple match, could be extended
+                        )
+                    )
                 else:
-                    conditions.append(FieldCondition(
-                        key=field,
-                        match=MatchValue(value=value)
-                    ))
+                    conditions.append(FieldCondition(key=field, match=MatchValue(value=value)))
 
             if conditions:
                 search_filters = Filter(must=conditions)
@@ -335,7 +332,7 @@ class QdrantIndexer:
                     collection_name=collection_name,
                     query_vector=query_embedding,
                     limit=limit,
-                    query_filter=search_filters
+                    query_filter=search_filters,
                 )
 
                 all_results[collection_name] = results
@@ -351,10 +348,10 @@ class QdrantIndexer:
                 for result in results:
                     table.add_row(
                         f"{result.score:.4f}",
-                        result.payload.get('qualified_name', result.payload.get('name', 'N/A')),
-                        result.payload.get('type', 'N/A'),
-                        result.payload.get('file_path', 'N/A')[:35],
-                        result.payload.get('embedding_quality', 'N/A')
+                        result.payload.get("qualified_name", result.payload.get("name", "N/A")),
+                        result.payload.get("type", "N/A"),
+                        result.payload.get("file_path", "N/A")[:35],
+                        result.payload.get("embedding_quality", "N/A"),
                     )
 
                 console.print(table)
@@ -369,22 +366,22 @@ class QdrantIndexer:
         """Test search across all collections (backward compatibility method)"""
         return self.advanced_search(query, limit)
 
-    def get_collection_stats(self) -> Dict[str, Dict]:
+    def get_collection_stats(self) -> dict[str, dict]:
         """Get statistics for all collections"""
         stats = {}
         for collection_name in self.collections.keys():
             try:
                 collection_info = self.client.get_collection(collection_name)
                 stats[collection_name] = {
-                    'vectors_count': collection_info.vectors_count,
-                    'status': str(collection_info.status),
-                    'config': {
-                        'vector_size': collection_info.config.params.vectors.size,
-                        'distance': str(collection_info.config.params.vectors.distance)
-                    }
+                    "vectors_count": collection_info.vectors_count,
+                    "status": str(collection_info.status),
+                    "config": {
+                        "vector_size": collection_info.config.params.vectors.size,
+                        "distance": str(collection_info.config.params.vectors.distance),
+                    },
                 }
             except Exception as e:
-                stats[collection_name] = {'error': str(e)}
+                stats[collection_name] = {"error": str(e)}
 
         return stats
 
@@ -423,36 +420,24 @@ class QdrantIndexer_2:
                     ),
                     sparse_vectors_config={
                         # Sparse identifiers (variable/function/class names)
-                        "symbols": SparseVectorParams(
-                            index=SparseIndexParams(on_disk=False)
-                        ),
+                        "symbols": SparseVectorParams(index=SparseIndexParams(on_disk=False)),
                         # BM25-like text search (docstrings, comments)
-                        "bm25": SparseVectorParams(
-                            index=SparseIndexParams(on_disk=False)
-                        ),
-                        "text-sparse": SparseVectorParams(
-                            index=SparseIndexParams(on_disk=False)
-                        ),
-                    }
+                        "bm25": SparseVectorParams(index=SparseIndexParams(on_disk=False)),
+                        "text-sparse": SparseVectorParams(index=SparseIndexParams(on_disk=False)),
+                    },
                 )
 
                 # Create payload indexes for filtering
                 if self.config.enable_payload_index:
                     # Index key fields
                     self.client.create_payload_index(
-                        collection_name=collection_name,
-                        field_name="language",
-                        field_schema=PayloadSchemaType.KEYWORD
+                        collection_name=collection_name, field_name="language", field_schema=PayloadSchemaType.KEYWORD
                     )
                     self.client.create_payload_index(
-                        collection_name=collection_name,
-                        field_name="type",
-                        field_schema=PayloadSchemaType.KEYWORD
+                        collection_name=collection_name, field_name="type", field_schema=PayloadSchemaType.KEYWORD
                     )
                     self.client.create_payload_index(
-                        collection_name=collection_name,
-                        field_name="file_path",
-                        field_schema=PayloadSchemaType.KEYWORD
+                        collection_name=collection_name, field_name="file_path", field_schema=PayloadSchemaType.KEYWORD
                     )
                     self.client.create_payload_index(
                         collection_name=collection_name,
@@ -465,49 +450,48 @@ class QdrantIndexer_2:
 
             self.collections[collection_name] = True
 
-    def _get_collection_for_chunk(self, chunk: Dict) -> str:
+    def _get_collection_for_chunk(self, chunk: dict) -> str:
         """Determine which collection a chunk should go into"""
-        chunk_type = chunk.get('type', 'unknown')
+        chunk_type = chunk.get("type", "unknown")
 
-        if chunk_type in ['function', 'method']:
+        if chunk_type in ["function", "method"]:
             return f"{self.config.collection_prefix}_functions"
-        elif chunk_type == 'class':
+        if chunk_type == "class":
             return f"{self.config.collection_prefix}_classes"
-        else:
-            return f"{self.config.collection_prefix}_modules"
+        return f"{self.config.collection_prefix}_modules"
 
-    def _prepare_payload(self, chunk: Dict) -> Dict:
+    def _prepare_payload(self, chunk: dict) -> dict:
         """Prepare chunk payload for Qdrant (remove embedding, keep metadata)"""
         payload = chunk.copy()
 
         # Remove embedding (stored separately)
-        payload.pop('embedding', None)
-        payload.pop('embedding_text', None)  # Don't need full text in payload
+        payload.pop("embedding", None)
+        payload.pop("embedding_text", None)  # Don't need full text in payload
 
         # Keep essential fields easily filterable
         essential = {
-            'id': chunk['id'],
-            'name': chunk['name'],
-            'type': chunk['type'],
-            'language': chunk['language'],
-            'file_path': chunk['file_path'],
-            'qualified_name': chunk.get('qualified_name', ''),
-            'code': chunk['code'],  # Keep original code
-            'start_line': chunk['start_line'],
-            'end_line': chunk['end_line'],
-            'complexity': chunk.get('complexity', 0),
-            'docstring': chunk.get('docstring'),
-            'signature': chunk.get('signature'),
-            'dependencies': chunk.get('dependencies', []),
-            'context': chunk.get('context', {}),
-            'metadata': chunk.get('metadata', {}),
-            'relationships': chunk.get('relationships', {}),
-            'analysis': chunk.get('analysis', {}),
+            "id": chunk["id"],
+            "name": chunk["name"],
+            "type": chunk["type"],
+            "language": chunk["language"],
+            "file_path": chunk["file_path"],
+            "qualified_name": chunk.get("qualified_name", ""),
+            "code": chunk["code"],  # Keep original code
+            "start_line": chunk["start_line"],
+            "end_line": chunk["end_line"],
+            "complexity": chunk.get("complexity", 0),
+            "docstring": chunk.get("docstring"),
+            "signature": chunk.get("signature"),
+            "dependencies": chunk.get("dependencies", []),
+            "context": chunk.get("context", {}),
+            "metadata": chunk.get("metadata", {}),
+            "relationships": chunk.get("relationships", {}),
+            "analysis": chunk.get("analysis", {}),
         }
 
         return essential
 
-    def index_chunks(self, chunks: List[Dict], batch_size: int = 100):
+    def index_chunks(self, chunks: list[dict], batch_size: int = 100):
         """Index chunks in Qdrant"""
         console.print(f"[cyan]Indexing {len(chunks)} chunks in Qdrant...[/cyan]")
 
@@ -529,28 +513,22 @@ class QdrantIndexer_2:
             for chunk in coll_chunks:
                 point = PointStruct(
                     id=int(chunk["id"], 16),  # Use chunk ID
-                    vector=chunk['embedding'],
-                    payload=self._prepare_payload(chunk)
+                    vector=chunk["embedding"],
+                    payload=self._prepare_payload(chunk),
                 )
                 points.append(point)
 
             # Batch upload
             for i in tqdm(range(0, len(points), batch_size), desc=f"Uploading to {collection_name}"):
-                batch = points[i:i + batch_size]
-                self.client.upsert(
-                    collection_name=collection_name,
-                    points=batch
-                )
+                batch = points[i : i + batch_size]
+                self.client.upsert(collection_name=collection_name, points=batch)
                 total_indexed += len(batch)
 
         console.print(f"[green]✓ Indexed {total_indexed} chunks successfully![/green]")
 
     def test_search(self, query: str, limit: int = 5):
         """Test search across all collections"""
-        console.print(Panel.fit(
-            f"[bold cyan]Test Search: '{query}'[/bold cyan]",
-            border_style="cyan"
-        ))
+        console.print(Panel.fit(f"[bold cyan]Test Search: '{query}'[/bold cyan]", border_style="cyan"))
 
         # Generate query embedding
         embedding_gen = EmbeddingGenerator(EmbeddingConfig())
@@ -564,11 +542,7 @@ class QdrantIndexer_2:
         for collection_name in self.collections.keys():
             console.print(f"\n[yellow]Results from {collection_name}:[/yellow]")
 
-            results = self.client.search(
-                collection_name=collection_name,
-                query_vector=query_embedding,
-                limit=limit
-            )
+            results = self.client.search(collection_name=collection_name, query_vector=query_embedding, limit=limit)
 
             table = Table(show_header=True, header_style="bold magenta")
             table.add_column("Score", width=8)
@@ -579,23 +553,21 @@ class QdrantIndexer_2:
             for result in results:
                 table.add_row(
                     f"{result.score:.4f}",
-                    result.payload.get('qualified_name', result.payload.get('name', 'N/A')),
-                    result.payload.get('type', 'N/A'),
-                    result.payload.get('file_path', 'N/A')
+                    result.payload.get("qualified_name", result.payload.get("name", "N/A")),
+                    result.payload.get("type", "N/A"),
+                    result.payload.get("file_path", "N/A"),
                 )
 
             console.print(table)
+
+
 def index_from_embedded_json(
-    json_path: str,
-    embedding_dim: int = 768,
-    collection_prefix: Optional[str] = None,
-    batch_size: Optional[int] = None
+    json_path: str, embedding_dim: int = 768, collection_prefix: str | None = None, batch_size: int | None = None
 ) -> None:
     """
     Load pre-embedded chunks from JSON file and index them into Qdrant collections.
     Adapted from old_code/embedder.py logic.
     """
-    from pathlib import Path
 
     path = Path(json_path)
     if not path.exists():
@@ -603,9 +575,9 @@ def index_from_embedded_json(
         return
 
     console.print(f"[cyan]Loading embedded chunks from {json_path}...[/cyan]")
-    with open(path, 'r') as f:
+    with Path(path).open() as f:
         data = json.load(f)
-        chunks = data.get('chunks', [])
+        chunks = data.get("chunks", [])
     console.print(f"[green]Loaded {len(chunks)} chunks[/green]")
 
     config = QdrantConfig()
@@ -620,7 +592,4 @@ def index_from_embedded_json(
     indexer.create_collections(embedding_dim)
     indexer.index_chunks(chunks, batch_size=batch_size)
 
-    console.print(Panel.fit(
-        "[bold green]✓ Indexing from embedded JSON complete![/bold green]",
-        border_style="green"
-    ))
+    console.print(Panel.fit("[bold green]✓ Indexing from embedded JSON complete![/bold green]", border_style="green"))
