@@ -1,35 +1,36 @@
-
 """
 Main package exports - Enhanced core components from Phase 2 merge
 """
 
 # Configuration classes (Phase 1)
-from .config import EmbeddingConfig, QdrantConfig
-
-# Enhanced preprocessing components (Phase 2)
-from .preprocessing import CodeChunk, ChunkPreprocessor, ChunkPreprocessor_2
+from .config import EmbeddingConfig
+from .config import QdrantConfig
 
 # Enhanced embedding components (Phase 2)
-from .embedding import EmbeddingGenerator, EmbeddingGenerator_2
+from .embedding import EmbeddingGenerator
+from .embedding import EmbeddingGenerator_2
+from .preprocessing import ChunkPreprocessor
+from .preprocessing import ChunkPreprocessor_2
+
+# Enhanced preprocessing components (Phase 2)
+from .preprocessing import CodeChunk
 
 # Enhanced retrieval components (Phase 2)
-from .retrieval import QdrantIndexer, QdrantIndexer_2
+from .retrieval import QdrantIndexer
+from .retrieval import QdrantIndexer_2
 
 __all__ = [
     # Configuration
-    'EmbeddingConfig',
-    'QdrantConfig',
-
+    "EmbeddingConfig",
+    "QdrantConfig",
     # Preprocessing
-    'CodeChunk',
-    'ChunkPreprocessor',
-    'ChunkPreprocessor_2',
-
+    "CodeChunk",
+    "ChunkPreprocessor",
+    "ChunkPreprocessor_2",
     # Embedding
-    'EmbeddingGenerator',
-    'EmbeddingGenerator_2',
-
+    "EmbeddingGenerator",
+    "EmbeddingGenerator_2",
     # Retrieval
-    'QdrantIndexer',
-    'QdrantIndexer_2'
+    "QdrantIndexer",
+    "QdrantIndexer_2",
 ]
