@@ -2,46 +2,46 @@
 Language configuration and mappings for code parsing
 """
 
-from tree_sitter import Language
-import tree_sitter_python as tspython
-import tree_sitter_javascript as tsjsvascript
-import tree_sitter_html as tshtml
 import tree_sitter_css as tscss
+import tree_sitter_html as tshtml
+import tree_sitter_javascript as tsjsvascript
+import tree_sitter_python as tspython
+from tree_sitter import Language
 
 
 class LanguageConfig:
     """Configuration for different programming languages"""
 
     LANGUAGE_MAP = {
-        '.py': 'python',
-        '.js': 'javascript',
-        '.mjs': 'javascript',
-        '.css': 'css',
-        '.html': 'html',
-        '.java': 'java',
-        '.cpp': 'cpp',
-        '.cc': 'cpp',
-        '.cxx': 'cpp',
-        '.c': 'c',
-        '.h': 'c',
-        '.hpp': 'cpp',
-        '.go': 'go',
-        '.rs': 'rust',
-        '.rb': 'ruby',
-        '.php': 'php',
-        '.cs': 'c_sharp',
+        ".py": "python",
+        ".js": "javascript",
+        ".mjs": "javascript",
+        ".css": "css",
+        ".html": "html",
+        ".java": "java",
+        ".cpp": "cpp",
+        ".cc": "cpp",
+        ".cxx": "cpp",
+        ".c": "c",
+        ".h": "c",
+        ".hpp": "cpp",
+        ".go": "go",
+        ".rs": "rust",
+        ".rb": "ruby",
+        ".php": "php",
+        ".cs": "c_sharp",
     }
 
     LANGUAGES = {
-        'python': Language(tspython.language()),
-        'javascript': Language(tsjsvascript.language()),
-        'html': Language(tshtml.language()),
-        'css': Language(tscss.language())
+        "python": Language(tspython.language()),
+        "javascript": Language(tsjsvascript.language()),
+        "html": Language(tshtml.language()),
+        "css": Language(tscss.language()),
     }
 
     # Tree-sitter query patterns for extracting definitions
     QUERIES = {
-        'python': """
+        "python": """
             (function_definition
                 name: (identifier) @func.name
                 parameters: (parameters) @func.params
@@ -51,7 +51,7 @@ class LanguageConfig:
                 name: (identifier) @class.name
                 body: (block) @class.body) @class
         """,
-        'javascript': """
+        "javascript": """
             (function_declaration
                 name: (identifier) @func.name
                 parameters: (formal_parameters) @func.params
@@ -69,21 +69,19 @@ class LanguageConfig:
     }
 
     DEFAULT_IGNORE_PATTERNS = [
-        '.git',
-        '__pycache__',
-        'node_modules',
-        'venv',
-        'env',
-        '.tinv'
-        '.venv',
-        'build',
-        'dist',
-        '*.pyc',
-        '*.pyo',
-        '*.egg-info',
-        '.pytest_cache',
-        '.mypy_cache',
-        '.ruff_cache'
-        'coverage',
-        '.coverage',
+        ".git",
+        "__pycache__",
+        "node_modules",
+        "venv",
+        "env",
+        ".tinv.venv",
+        "build",
+        "dist",
+        "*.pyc",
+        "*.pyo",
+        "*.egg-info",
+        ".pytest_cache",
+        ".mypy_cache",
+        ".ruff_cachecoverage",
+        ".coverage",
     ]
