@@ -1,17 +1,9 @@
-
 """
 Retrieval module exports - Enhanced with advanced indexing and search capabilities
 """
 
-from .search import (
-    QdrantIndexer,  # Enhanced version with advanced management
-    QdrantIndexer_2  # Original version for backward compatibility
-)
-
 from .rag_system import CodeRAG_2
+from .search import QdrantIndexer  # Enhanced version with advanced management
+from .search import QdrantIndexer_2  # Original version for backward compatibility
 
-__all__ = [
-    'QdrantIndexer',
-    'QdrantIndexer_2',
-    'CodeRAG_2'
-]
+__all__ = ["CodeRAG_2", "QdrantIndexer", "QdrantIndexer_2"]
