@@ -1,11 +1,14 @@
-import pytest
 import os
+
+import pytest
 from pydantic import ValidationError
-from src.config.embedding_config import EmbeddingConfig
-from src.config.qdrant_config import QdrantConfig
 from qdrant_client.models import Distance
 
+from src.config.embedding_config import EmbeddingConfig
+from src.config.qdrant_config import QdrantConfig
+
 # --- Tests for EmbeddingConfig ---
+
 
 def test_embedding_config_defaults():
     config = EmbeddingConfig(_env_file=None)
@@ -17,6 +20,7 @@ def test_embedding_config_defaults():
     assert config.timeout == 30
     assert config.embedding_model_name == "ai/embeddinggemma"
 
+
 def test_embedding_config_env_vars(monkeypatch):
     monkeypatch.setenv("model_url", "http://custom-host:8000")
     monkeypatch.setenv("model_name", "custom-model")
@@ -24,7 +28,7 @@ def test_embedding_config_env_vars(monkeypatch):
     monkeypatch.setenv("batch_size", "64")
     monkeypatch.setenv("max_retries", "5")
     monkeypatch.setenv("timeout", "60")
-    
+
     config = EmbeddingConfig(_env_file=None)
     assert config.model_url == "http://custom-host:8000"
     assert config.model_name == "custom-model"
@@ -34,33 +38,40 @@ def test_embedding_config_env_vars(monkeypatch):
     assert config.timeout == 60
     assert config.embedding_model_name == "custom-model"
 
+
 def test_embedding_config_init_backward_compatibility_model_name_only():
     config = EmbeddingConfig(model_name="new-model", _env_file=None)
     assert config.model_name == "new-model"
     assert config.embedding_model_name == "new-model"
+
 
 def test_embedding_config_init_backward_compatibility_embedding_model_name_only():
     config = EmbeddingConfig(embedding_model_name="old-model", _env_file=None)
     assert config.model_name == "old-model"
     assert config.embedding_model_name == "old-model"
 
+
 def test_embedding_config_init_backward_compatibility_both_provided():
     config = EmbeddingConfig(model_name="new-model", embedding_model_name="old-model-ignored", _env_file=None)
     assert config.model_name == "new-model"
     assert config.embedding_model_name == "old-model-ignored"
 
+
 def test_embedding_config_effective_model_name():
     config = EmbeddingConfig(model_name="primary-model", embedding_model_name="fallback-model", _env_file=None)
     assert config.effective_model_name == "primary-model"
-    
+
     config = EmbeddingConfig(model_name="", embedding_model_name="fallback-model", _env_file=None)
     assert config.effective_model_name == "fallback-model"
+
 
 def test_embedding_config_get_api_endpoint():
     config = EmbeddingConfig(model_url="http://test.com:1234", _env_file=None)
     assert config.get_api_endpoint() == "http://test.com:1234/embeddings"
 
+
 # --- Tests for QdrantConfig ---
+
 
 def test_qdrant_config_defaults():
     config = QdrantConfig(_env_file=None)
@@ -74,6 +85,7 @@ def test_qdrant_config_defaults():
     assert config.batch_size == 100
     assert config.on_disk_vectors is False
     assert config.on_disk_sparse_vectors is False
+
 
 def test_qdrant_config_env_vars(monkeypatch):
     monkeypatch.setenv("host", "qdrant.cloud")
@@ -99,26 +111,32 @@ def test_qdrant_config_env_vars(monkeypatch):
     assert config.on_disk_vectors is True
     assert config.on_disk_sparse_vectors is True
 
+
 def test_qdrant_config_get_collection_names():
     config = QdrantConfig(collection_prefix="my_project", _env_file=None)
     expected_names = ["my_project_functions", "my_project_classes", "my_project_modules"]
     assert config.get_collection_names() == expected_names
 
+
 def test_qdrant_config_get_connection_url_no_api_key():
     config = QdrantConfig(host="test-host", port=1234, qdrant_api_key="", _env_file=None)
     assert config.get_connection_url() == "http://test-host:1234"
+
 
 def test_qdrant_config_get_connection_url_with_api_key():
     config = QdrantConfig(host="cloud-host", port=5678, qdrant_api_key="some-key", _env_file=None)
     assert config.get_connection_url() == "https://cloud-host:5678"
 
+
 def test_qdrant_config_get_client_config_no_api_key():
     config = QdrantConfig(host="local", port=1111, qdrant_api_key="", _env_file=None)
     assert config.get_client_config() == {"host": "local", "port": 1111}
 
+
 def test_qdrant_config_get_client_config_with_api_key():
     config = QdrantConfig(host="remote", port=2222, qdrant_api_key="remote-key", _env_file=None)
     assert config.get_client_config() == {"host": "remote", "port": 2222, "api_key": "remote-key"}
+
 
 def test_qdrant_config_distance_metric_invalid_value(monkeypatch):
     monkeypatch.setenv("distance_metric", "INVALID_METRIC")
